更改 aid cid 数据类型 `int` -> `long`

+++ app/src/main/kotlin/dev/aaa1115910/bv/activities/video/VideoPlayerActivity.kt
@@ -28,8 +28,8 @@ class VideoPlayerActivity : ComponentActivity() {
         private val logger = KotlinLogging.logger { }
         fun actionStart(
             context: Context,
-            avid: Int,
-            cid: Int,
+            avid: Long,
+            cid: Long,
             title: String,
             partTitle: String,
             played: Int,
@@ -70,8 +70,8 @@ class VideoPlayerActivity : ComponentActivity() {
         playerViewModel.preparePlayer(player)
 
         if (intent.hasExtra("avid")) {
-            val aid = intent.getIntExtra("avid", 170001)
-            val cid = intent.getIntExtra("cid", 170001)
+            val aid = intent.getLongExtra("avid", 170001)
+            val cid = intent.getLongExtra("cid", 170001)
             val title = intent.getStringExtra("title") ?: "Unknown Title"
             val partTitle = intent.getStringExtra("partTitle") ?: "Unknown Part Title"
             val played = intent.getIntExtra("played", 0)
