+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/entity/video/RelatedVideo.kt
@@ -4,7 +4,7 @@ import bilibili.app.view.v1.authorOrNull
 import dev.aaa1115910.biliapi.entity.user.Author
 
 data class RelatedVideo(
-    val aid: Int,
+    val aid: Long,
     val cover: String,
     val title: String,
     val duration: Int,
@@ -16,7 +16,7 @@ data class RelatedVideo(
 ) {
     companion object {
         fun fromRelate(relate: bilibili.app.view.v1.Relate) = RelatedVideo(
-            aid = relate.aid.toInt(),
+            aid = relate.aid,
             cover = relate.pic,
             title = relate.title,
             duration = relate.duration.toInt(),
